refactor: extract cache utils to module, remove dead center_padding code
- Move _CachedSample and _clone_for_cache to new cache.py module
  (data/augmentation/cache.py) for cleaner separation of concerns
- Remove center_padding parameter from Resize transform (unused in any
  recipe or runtime path; will be reintroduced in Ultralytics PR)
- Remove center_padding property from OVModel (dead code, never consumed
  by the auto-configurator or evaluation pipeline)
- Remove corresponding center_padding tests
- Update test imports to use new cache module path
